overflow fix

+++ lib/utils.h
@@ -4,7 +4,7 @@
 #define UTILS
 
 #define TIME_STR_SIZE	64
-#define SEC_70 			(70*365+17)*24*60*60
+#define SEC_70 			(uint32_t)(70*365+17)*24*60*60
 #define CONVERT70_00(x) x+SEC_70
 
 #endif
